requirements.txt
Django
psycopg2-binary>=2.8
djangorestframework
django-storages==1.9.1
PyYAML
django-cors-headers
django-url-filter==0.3.15